Remove link_directory and move stuff to asset_list
